move page to Join folder, add octopus image and go to top
